Add functions to extract border/center from image
The borders are extracted as 16x48 or 48x16 patches clockwise from the
top left, flattened, then concatenated. The center is the 32x32 patch
not part of the borders, and can be flattened. Using the fully
flattened configuration, this can be used to provide input to an
densely connected regular MLP.